+++ test/tag.tsx
@@ -1,12 +1,4 @@
 /** @jsx JSXSlack.h */
-import {
-  Dialog,
-  Input,
-  Option as DialogOption,
-  Select as DialogSelect,
-  Textarea,
-  UsersSelect,
-} from '../src/dialog'
 import JSXSlack, {
   Actions,
   Blocks,
@@ -21,12 +13,6 @@ import JSXSlack, {
 } from '../src/index'
 
 describe('Tagged template', () => {
-  let warnSpy: jest.SpyInstance
-
-  beforeEach(() => {
-    warnSpy = jest.spyOn(console, 'warn')
-  })
-
   it('allows converting Block Kit JSX to JSON without transpiler', () => {
     const count = 2
     const template = jsxslack`
@@ -72,57 +58,6 @@ describe('Tagged template', () => {
         </Blocks>
       )
     )
-    expect(warnSpy).toBeCalledTimes(0)
-  })
-
-  it('allows converting dialog JSX to JSON without transpiler', () => {
-    const template = jsxslack`
-      <Dialog callbackId="callback" title="test">
-        <Input type="hidden" name="hiddenState" value=${['a', 'b', 'c']} />
-        <Input type="text" name="foo" label="foo" required />
-        <Textarea name="bar" label="bar" />
-        <Select name="select" label="select">
-          <Option value="1">one</Option>
-          <Option value="2">two</Option>
-          <Option value="3">three</Option>
-        </Select>
-        <Input type="submit" value="Submit dialog" />
-      </Dialog>
-    `
-
-    expect(template).toStrictEqual(
-      JSXSlack(
-        <Dialog callbackId="callback" title="test">
-          <Input type="hidden" name="hiddenState" value={['a', 'b', 'c']} />
-          <Input type="text" name="foo" label="foo" required />
-          <Textarea name="bar" label="bar" />
-          <DialogSelect name="select" label="select">
-            <DialogOption value="1">one</DialogOption>
-            <DialogOption value="2">two</DialogOption>
-            <DialogOption value="3">three</DialogOption>
-          </DialogSelect>
-          <Input type="submit" value="Submit dialog" />
-        </Dialog>
-      )
-    )
-    expect(warnSpy).toBeCalledTimes(1)
-  })
-
-  it('can use imported components through interpolation', () => {
-    expect(jsxslack`
-      <${Dialog} callbackId="callback" title="test">
-        <${Input} type="hidden" name="hiddenState" value=${['a', 'b', 'c']} />
-        <${Input} type="text" name="foo" label="foo" required />
-        <${Textarea} name="bar" label="bar" />
-        <${DialogSelect} name="select" label="select">
-          <${DialogOption} value="1">one<//>
-          <${DialogOption} value="2">two<//>
-          <${DialogOption} value="3">three<//>
-        <//>
-        <${Input} type="submit" value="Submit dialog" />
-      <//>
-    `).toMatchSnapshot()
-    expect(warnSpy).toBeCalledTimes(1)
   })
 
   it('can use fragmented options in <Select>', () => {
@@ -152,7 +87,6 @@ describe('Tagged template', () => {
         </Blocks>
       )
     )
-    expect(warnSpy).toBeCalledTimes(0)
   })
 
   it('can use interpolations through conditional rendering', () => {
@@ -172,7 +106,6 @@ describe('Tagged template', () => {
         </Blocks>
       )
     )
-    expect(warnSpy).toBeCalledTimes(0)
   })
 
   it('has same decode behavior compatible with JSX for HTML entities', () => {
@@ -265,35 +198,6 @@ describe('Tagged template', () => {
       expect(jsxslack.fragment`<${Component}>test<//>`).toStrictEqual(
         func('test')
       )
-    })
-
-    it('allows using "Dialog." prefix to use select components for dialog', () => {
-      const Component = ({ value }) => jsxslack.fragment`
-        <Dialog.UsersSelect name="first-user" label="First user" initialUser=${value} />
-        <Dialog.UsersSelect name="second-user" label="Second user" initialUser=${value} />
-      `
-
-      expect(
-        jsxslack`<Dialog callbackId="dialog" title="Dialog"><${Component} value="U01234567" /></Dialog>`
-      ).toStrictEqual(
-        JSXSlack(
-          <Dialog callbackId="dialog" title="Dialog">
-            <UsersSelect
-              name="first-user"
-              label="First user"
-              initialUser="U01234567"
-            />
-            <UsersSelect
-              name="second-user"
-              label="Second user"
-              initialUser="U01234567"
-            />
-          </Dialog>
-        )
-      )
-
-      // Warn 3 times by each components
-      expect(warnSpy).toBeCalledTimes(3)
     })
   })
 })